requirements.txt
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
filelock==3.8.2
huggingface-hub==0.11.1
idna==3.4
leveldb==0.201
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-contrib-python==4.6.0.66
packaging==22.0
Pillow==9.3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scenedetect==0.6.1
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typing-extensions==4.4.0
urllib3==1.26.13
